ratings fixed

+++ db/migrate/20150117084450_add_lat_lng_to_places.rb
@@ -2,7 +2,7 @@ class AddLatLngToPlaces < ActiveRecord::Migration
   def change
   	add_column :places, :lat, :decimal
   	add_column :places, :lng, :decimal
-  	add_column :places, :rating, :integer
+  	add_column :places, :rating, :float
   	add_column :places, :icon_url, :string
   	add_column :places, :address, :string
   end
